rename: 파일명 변경
폴더와 폴더 안의 이름이 겹쳐 파일 이름을 index.ts로 변경함
