+++ src/views/home/index.ts
@@ -1,6 +1,6 @@
 import Layout from "@comp/Layout";
 import styles from "./index.css";
-import { createRotatingText } from "@/util/rotatingText/rotatingText";
+import { createRotatingText } from "@/util/rotatingText";
 
 export default function Home() {
   const rotatingWords = ["송영범", "나나나", "오오오", "징징징"];
